# Fix #124: Treat :character offsets as UTF-16 code units