https://microsoft.github.io/language-server-protocol/specification#text-documents

The spec says that :character offsets are based on a UTF-16
representation.  Previously we were assuming that :character specifies
the number of characters.

* eglot.el (eglot--count-utf-16-code-units): New function.
(eglot--pos-to-lsp-position):
(eglot--lsp-position-to-point): Use it.