@MaskRay 2018-10-03
Rather than use UTF-16 unconditionally, you may consider hide this under a flag. Some other language clients don't UTF-8 (LanguageClient-neovim, lsp-mode) either and this isn't an issue in practice because code points in Supplementary Planes are rarely used.

There is a bug tracking the conversion to code points and you may raise that topic again https://github.com/Microsoft/language-server-protocol/issues/376