@mkcms 2018-11-07
@MaskRay 

> Please provide the UTF-16 transcoding under a `defvar-local` (or whatever) toggle.

I pushed a commit which adds a `defvar` which controls whether UTF-16 encoding is used.