@joaotavora 2018-11-07
@mkcms, have a look at https://github.com/joaotavora/eglot/commit/1999245aeeeba99a02b8a5101860cf78acd384f6 and tell me what you think. There's a very simple test, which we could overhaul with many difficult situations.

It'd be interesting to compare the performance of your approach and mine's (which I think uses more of Emacs's built-in machinery to do what you do by hand, and is therefore more readable).

Interestingly, Emacs's `bufferpos-to-filepos` can return `exact` or `approximate` char positions, which supposedly affects performance.  I'm using `exact` but we could use something else.