@joaotavora 2018-11-08
Actually look here (https://github.com/joaotavora/eglot/commit/3bbc93718dfb1610cf43774ef4dc2377e0f6448a), I bungled that other commit quite badly with some stuff that wasn't supposed to go in.