@mkcms 2018-11-08
@joaotavora Thanks for the review.

> * bad in performance, more than probably;

Yes, of course it will slow down the position conversion functions. What I meant before was overall performance that affects the user. My reasoning was that converting Emacs to LSP positions only takes a fraction of the overall time (e.g. when applying workspace edits, I/O and buffer modifications take the most time). Obviously, I haven't really benchmarked, so it's possible that my change would slow down Eglot by a large factor in other use cases. I only tested this change by doing regular coding work and I haven't noticed any slowdown. 

> * questionable in benefit (though I have understood your #124 bug report, and it is indeed a bug)

I agree that 99% of the time, this is just redundant computation. Actually this issue only happened to me exactly once, and it wasn't even serious.

---

> Actually look here ([3bbc937](https://github.com/joaotavora/eglot/commit/3bbc93718dfb1610cf43774ef4dc2377e0f6448a)),

I also looked at the `bufferpos`/`filepos` functions before. I didn't use them because they always calculate the position from beginning of buffer. This seems like significant overhead, although I have no idea how they're implemented. Maybe they're heavily optimized.

---

Maybe instead of unconditionally converting the positions, perhaps we could add a `eglot-position-conversion-function` (or just agree to keep `eglot--lsp-position-to-point` and `eglot--pos-to-lsp-position` always backwards-compatible) so that users who experience this problem can override those functions, and add a note in README.