@mkcms 2018-11-11
> This is a good idea.

Alright, I will do this.

---

I also did some benchmarks, comparing execution times of `eglot--lsp-position-to-point` on master, this branch and scratch/fix-124-idea.
The test case is converting a couple of thousand lsp positions to emacs points in a 1000 line file with some wide characters.
Here are the results:

```
$ ./run-benchmark.sh 
master 1.4869308209999998s
fix/issue-124 1.803970126s
scratch/fix-124-idea 7.018219207s
```

I've also managed to optimize my solution to get these times:

```
$ ./run-benchmark.sh 
master 1.389679567s
fix/issue-124 1.4014020470000002s
scratch/fix-124-idea 6.454922788999999s
```

The benchmark suite: [benchmark.zip](https://github.com/joaotavora/eglot/files/2569383/benchmark.zip)