@joaotavora 2018-11-11
@mkcms, I'll have a look at your benchmark but in the meantime, to avoid confusion of mental translations I've pushed two branches `scratch/fix-124-joaotavora` and `scratch/fix-125-mkcms` to hold the two approaches. I'll push tweaks to the first while you push to the second, and in the end we can decide which is the best.