@joaotavora 2018-11-11
@mkcms your benchmark idea is very good,  but I don't understand `insert-file-contents-literally`.  Shouldn't we be opening the test file normally, with its utf-8 encoding?

Also, `positions.el` as returned from an hypothetical UTF-16-counting LSP server doesn't point to the positions of the multibyte characters in the generated `text.txt` file.  Please verify on your side.