@mkcms 2018-11-11
> I don't understand `insert-file-contents-literally`. Shouldn't we be opening the test file normally, with its utf-8 encoding?

Right, I missed that.

Attached is a new benchmark suite which will also test the number of times each solution was incorrect.
Here is my output for now:
```
$ ./run-benchmark.sh 
master 1.687929406s mismatches: 16960
scratch/fix-124-mkcms 2.620455026s mismatches: 0
scratch/fix-124-joaotavora 15.336856984s mismatches: 0
```

With optimization:
```
$ ./run-benchmark.sh 
master 1.655803003s mismatches: 17320
scratch/fix-124-mkcms 1.7465138349999998s mismatches: 0
scratch/fix-124-joaotavora 13.732144245s mismatches: 0
```

The optimization is this:
Instead of iterating from the beginning of line, first `forward-char` the same number of chars as `:character`, then iterate backwards until the number of code units between `point-at-bol` and `point` equals `:character`.

---
[benchmark.zip](https://github.com/joaotavora/eglot/files/2569662/benchmark.zip)